8/22-commit

+++ app/Http/Controllers/Tweet/DeleteController.php
@@ -5,12 +5,16 @@ namespace App\Http\Controllers\Tweet;
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use App\Models\Tweet;
-use App\Http\Requests\Tweet\CreateRequest;
+use App\Services\TweetService;
+use Symfony\Component\HttpKernel\Exception\AccessDeniedHttpException;
 
 class DeleteController extends Controller
 {
-    public function e(Request $request){
+    public function e(Request $request , TweetService $tweetService){
         $tweetId =(int) $request->route('tweetId');
+        if(!$tweetService->checkOwnTweet($request->user()->id, $tweetId)){
+            throw new AccessDeniedHttpException();
+        }
         $tweets = Tweet::where('id',$tweetId)->firstOrFail();
         $tweets->delete();
         return redirect()->route('tweet.index')->with('success','Tweet deleted');
